Distributed refactored; NumGridVSR implements parametrized.
